+++ include/nn/layers_impl/flatten_layer.hpp
@@ -19,14 +19,14 @@ namespace tnn {
 // Flatten Layer for converting from 4D to 2D tensors (for compatibility with dense layers)
 template <typename T = float> class FlattenLayer : public StatelessLayer<T> {
 private:
-  std::unordered_map<int, std::vector<size_t>> micro_batch_original_shapes_;
+  std::unordered_map<size_t, std::vector<size_t>> micro_batch_original_shapes_;
 
 public:
   explicit FlattenLayer(const std::string &name = "flatten");
 
-  Tensor<T> forward(const Tensor<T> &input, int micro_batch_id = 0) override;
+  Tensor<T> forward(const Tensor<T> &input, size_t micro_batch_id = 0) override;
   Tensor<T> backward(const Tensor<T> &grad_output,
-                     int micro_batch_id = 0) override;
+                     size_t micro_batch_id = 0) override;
 
   std::string type() const override;
   LayerConfig get_config() const override;
